ocr/requirements.txt
certifi==2018.10.15
mkl-fft==1.0.6
mkl-random==1.0.1
numpy==1.15.4
opencv-python==3.4.3.18
pandas==0.23.4
Pillow==5.3.0
pytesseract==0.2.5
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
